+++ src/eca/config.clj
@@ -45,14 +45,10 @@
                          :key "${env:OPENAI_API_KEY}"
                          :requiresAuth? true
                          :models {"gpt-5.2" {}
-                                  "gpt-5.1" {}
-                                  "gpt-5-codex" {}
-                                  "gpt-5" {}
+                                  "gpt-5.2-codex" {}
                                   "gpt-5-mini" {}
                                   "gpt-5-nano" {}
-                                  "gpt-4.1" {}
-                                  "o4-mini" {}
-                                  "o3" {}}}
+                                  "gpt-4.1" {}}}
                "anthropic" {:api "anthropic"
                             :url "${env:ANTHROPIC_API_URL:https://api.anthropic.com}"
                             :key "${env:ANTHROPIC_API_KEY}"