+++ src/eca/llm_providers/anthropic.clj
@@ -72,12 +72,13 @@
   (let [url (str api-url (or url-relative-path messages-path))
         reason-id (str (random-uuid))
         oauth? (= :auth/oauth auth-type)
-        headers (assoc-some
-                 {"anthropic-version" "2023-06-01"
-                  "Content-Type" "application/json"}
-                 "x-api-key" (when-not oauth? api-key)
-                 "Authorization" (when oauth? (str "Bearer " api-key))
-                 "anthropic-beta" (when oauth? "oauth-2025-04-20"))
+        headers (client/merge-llm-headers
+                 (assoc-some
+                  {"anthropic-version" "2023-06-01"
+                   "Content-Type" "application/json"}
+                  "x-api-key" (when-not oauth? api-key)
+                  "Authorization" (when oauth? (str "Bearer " api-key))
+                  "anthropic-beta" (when oauth? "oauth-2025-04-20")))
         response* (atom nil)
         on-error (if on-stream
                    on-error
@@ -363,9 +364,9 @@
   (swap! db* assoc-in [:auth provider] {:step :login/waiting-login-method})
   (send-msg! (multi-str "Now, inform the login method:"
                         ""
-                        "max: Claude Pro/Max"
-                        "console: Create API Key"
-                        "manual: Manually enter API Key")))
+                        "- max: Claude Pro/Max"
+                        "- console: Automatically create API Key and use it"
+                        "- manual: Manually enter API Key")))
 
 (defmethod f.login/login-step ["anthropic" :login/waiting-login-method] [{:keys [db* input provider send-msg!]}]
   (case input
@@ -374,14 +375,14 @@
       (swap! db* assoc-in [:auth provider] {:step :login/waiting-provider-code
                                             :mode :max
                                             :verifier verifier})
-      (send-msg! (format "Open your browser at `%s` and authenticate at Anthropic.\nThen paste the code generated in the chat and send it to continue the authentication."
+      (send-msg! (format "Open your browser at:\n\n%s\n\nAuthenticate at Anthropic, then paste the code generated in the chat and send it to continue the authentication."
                          url)))
     "console"
     (let [{:keys [verifier url]} (oauth-url :console)]
       (swap! db* assoc-in [:auth provider] {:step :login/waiting-provider-code
                                             :mode :console
                                             :verifier verifier})
-      (send-msg! (format "Open your browser at `%s` and authenticate at Anthropic.\nThen paste the code generated in the chat and send it to continue the authentication."
+      (send-msg! (format "Open your browser at:\n\n%s\n\nAuthenticate at Anthropic, then paste the code generated in the chat and send it to continue the authentication."
                          url)))
     "manual"
     (do